pyproject.toml
[project]
name = "smatr_tg_bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Mileshkin Oleksandr",email = "mileshkin89@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "openai (>=1.68.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-telegram-bot (>=22.0,<23.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-docx (>=1.1.2)",
    "reportlab (>=4.3.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
